outputs/jsr-main/backend/requirements.txt
flask==3.0.0
pillow==10.2.0
numpy==1.26.3
requests==2.31.0
matplotlib==3.8.2